+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added initialization messages at the start of the flow
+### Changed
+- Query explainer: improved `INPUT DATA` section view
+  - added `Run` and `Copy` buttons for a sql code
 ### Fixed
 - Flows table: fixed broken filters in pagination